# Optimise `global in` over distributed table