### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
todo

---

``` sql
CREATE TABLE default.t
(
    `a` UInt64
)
ENGINE = MergeTree
ORDER BY tuple();

CREATE TABLE default.dt
(
    `a` UInt64
)
ENGINE = Distributed('test_cluster_two_shards', 'default', 't', rand());

INSERT INTO dt SELECT * FROM numbers_mt(1e8);

WITH filterIds AS
    (
        SELECT a
        FROM dt
        WHERE (a % 2) = 0
    )
SELECT a
FROM dt
WHERE a GLOBAL IN (filterIds)
GROUP BY a
FORMAT `Null`;
```

1e5: 0.021s vs 0.021s
1e6: 0.1s vs 0.22s
1e7: 1s vs 1.7s
1e8: 14s now vs 22s on master