@nickitat 2023-02-20
actually it is more of a workaround than a proper fix. let's merge if it doesn't introduce any degradation. I will dig into the problem later in background.