@nickitat 2023-02-21
actual difference turned out to be much smaller. because locally I have `memory_profiler_step = 1MB`.
for the query from description I see 12s vs 14.5s and it matches perf of the rewritten query that uses temp table.